make ModbusContext a trait and move it sample impl to ModbusStorage
