+++ examples/servers/tcp.rs
@@ -7,12 +7,12 @@ use lazy_static::lazy_static;
 use std::sync::RwLock;
 
 use rmodbus::{
-    server::{context::ModbusContextFull, ModbusFrame},
+    server::{context::ModbusContext, storage::ModbusStorageFull, ModbusFrame},
     ModbusFrameBuf, ModbusProto,
 };
 
 lazy_static! {
-    pub static ref CONTEXT: RwLock<ModbusContextFull> = RwLock::new(ModbusContextFull::new());
+    pub static ref CONTEXT: RwLock<ModbusStorageFull> = RwLock::new(ModbusStorageFull::new());
 }
 
 pub fn tcpserver(unit: u8, listen: &str) {
@@ -35,8 +35,8 @@ pub fn tcpserver(unit: u8, listen: &str) {
                 }
                 if frame.processing_required {
                     let result = match frame.readonly {
-                        true => frame.process_read(&CONTEXT.read().unwrap()),
-                        false => frame.process_write(&mut CONTEXT.write().unwrap()),
+                        true => frame.process_read(&*CONTEXT.read().unwrap()),
+                        false => frame.process_write(&mut *CONTEXT.write().unwrap()),
                     };
                     if result.is_err() {
                         println!("frame processing error");
