+++ src/server/mod.rs
@@ -1,4 +1,5 @@
 pub mod context;
+pub mod storage;
 
 use crate::consts::{
     MODBUS_ERROR_ILLEGAL_DATA_ADDRESS, MODBUS_ERROR_ILLEGAL_DATA_VALUE,
@@ -142,9 +143,9 @@ impl<'a, V: VectorTrait<u8>> ModbusFrame<'a, V> {
         }
     }
     /// Process write functions
-    pub fn process_write<const C: usize, const D: usize, const I: usize, const H: usize>(
+    pub fn process_write<C: context::ModbusContext>(
         &mut self,
-        ctx: &mut context::ModbusContext<C, D, I, H>,
+        ctx: &mut C,
     ) -> Result<(), ErrorKind> {
         match self.func {
             MODBUS_SET_COIL => {
@@ -216,10 +217,7 @@ impl<'a, V: VectorTrait<u8>> ModbusFrame<'a, V> {
         }
     }
     /// Process read functions
-    pub fn process_read<const C: usize, const D: usize, const I: usize, const H: usize>(
-        &mut self,
-        ctx: &context::ModbusContext<C, D, I, H>,
-    ) -> Result<(), ErrorKind> {
+    pub fn process_read<C: context::ModbusContext>(&mut self, ctx: &C) -> Result<(), ErrorKind> {
         match self.func {
             MODBUS_GET_COILS | MODBUS_GET_DISCRETES => {
                 // funcs 1 - 2
